# Use monotonically increasing time