--- a/lib/new_relic/agent/custom_event_aggregator.rb
+++ b/lib/new_relic/agent/custom_event_aggregator.rb
@@ -46,7 +46,8 @@ module NewRelic
 
       def create_event(type, priority, attributes)
         [
-          { TYPE => type, TIMESTAMP => Time.now.to_i,
+          { TYPE => type,
+            TIMESTAMP => Process.clock_gettime(Process::CLOCK_REALTIME).to_i,
             PRIORITY => priority
           },
           AttributeProcessing.flatten_and_coerce(attributes)