--- a/lib/new_relic/agent/distributed_tracing/distributed_trace_payload.rb
+++ b/lib/new_relic/agent/distributed_tracing/distributed_trace_payload.rb
@@ -50,7 +50,7 @@ module NewRelic
 
           payload.id = current_segment_id(transaction)
           payload.transaction_id = transaction.guid
-          payload.timestamp = (Process.clock_gettime(Process::CLOCK_REALTIME) * 1000).round
+          payload.timestamp = Process.clock_gettime(Process::CLOCK_REALTIME, :millisecond)
           payload.trace_id = transaction.trace_id
           payload.sampled = transaction.sampled?
           payload.priority = transaction.priority