--- a/lib/new_relic/agent/pipe_service.rb
+++ b/lib/new_relic/agent/pipe_service.rb
@@ -61,6 +61,9 @@ module NewRelic
         write_to_pipe(:sql_trace_data, sql) if sql
       end
 
+      # TODO: IS THE TIME ARGUMENT NECESSARY?
+      # IF NOT, REMOVE THE ARGUMENT FROM THE CALLS TO THIS METHOD
+      # OFTEN @service.shutdown()
       def shutdown(time)
         @pipe.close if @pipe
       end