--- a/lib/new_relic/agent/sql_sampler.rb
+++ b/lib/new_relic/agent/sql_sampler.rb
@@ -46,8 +46,7 @@ module NewRelic
           NewRelic::Agent::Database.should_record_sql?(:slow_sql)
       end
 
-      # TODO: IS THE START TIME ARGUMENT USED? COULD IT BE REMOVED?
-      def on_start_transaction(state, start_time, uri=nil)
+      def on_start_transaction(state, uri=nil)
         return unless enabled?
 
         state.sql_sampler_transaction_data = TransactionSqlData.new