… infinite_tracing/test/infinite_tracing/agent_integrations/datastore_segment_test.rb
… infinite_tracing/test/infinite_tracing/agent_integrations/external_request_segment_test.rb
… infinite_tracing/test/infinite_tracing/agent_integrations/segment_test.rb
… infinite_tracing/test/support/fake_trace_observer_helpers.rb
… lib/new_relic/agent/threading/backtrace_service.rb
… lib/new_relic/agent/transaction.rb
… lib/new_relic/agent/transaction/abstract_segment.rb
… lib/new_relic/agent/transaction_time_aggregator.rb
… lib/new_relic/agent/worker_loop.rb
… lib/new_relic/noticed_error.rb
… test/agent_helper.rb
… test/environments/lib/environments/runner.rb
… test/helpers/minitest.rb
… test/multiverse/suites/agent_only/adaptive_sampler_test.rb
… test/multiverse/suites/agent_only/audit_log_test.rb
… test/multiverse/suites/agent_only/custom_analytics_events_test.rb
… test/multiverse/suites/agent_only/custom_queue_time_test.rb
… test/multiverse/suites/agent_only/error_events_test.rb
… test/multiverse/suites/agent_only/exclusive_time_test.rb
… test/multiverse/suites/agent_only/harvest_timestamps_test.rb
… test/multiverse/suites/agent_only/key_transactions_test.rb
… test/multiverse/suites/agent_only/span_events_test.rb
… test/multiverse/suites/rack/example_app.rb
… test/multiverse/suites/rack/rack_auto_instrumentation_test.rb
… test/multiverse/suites/rails/app/views/views/_a_partial.html.erb
… test/multiverse/suites/rails/app/views/views/index.html.erb
… test/multiverse/suites/rails/gc_instrumentation_test.rb
… test/multiverse/suites/rails/queue_time_test.rb
… test/multiverse/suites/rails/view_instrumentation_test.rb
… test/new_relic/agent/adaptive_sampler_test.rb