_Before contributing, please read our [contributing guidelines](https://github.com/newrelic/newrelic-ruby-agent/blob/main/CONTRIBUTING.md) and [code of conduct](https://github.com/newrelic/.github/blob/master/CODE_OF_CONDUCT.md)._

# Overview
Replace locations where `Time.now` was coerced into floats or used to perform math operations with [Process.clock_gettime](https://ruby-doc.org/core-3.0.2/Process.html#method-c-clock_gettime) to improve performance and accuracy of time references.

`Process.clock_gettime(Process::CLOCK_REALTIME)`:
- Represents the wall clock time (ex: August 26, 2021 at 4:03 PM) 
- Most performant option in benchmarking tests
- Could still be subject to changes in system time (ex: manually changing a clock's time, leap seconds)
- Can be coerced into a `Time` object using `Time.at(Process.clock_gettime(Process::CLOCK_REALTIME)` 

`Process.clock_gettime(Process::CLOCK_MONOTONIC)`:
- Represents a continuously increasing timer that began with the life of the machine
- Within the margin of error on tests in speed
- Can't be converted into a human-readable timestamp
- Ideal for durations

`Time.now`:
- Creates a time object representing the current wall clock time (ex: August 26, 2021 at 4:03 PM) 
- Ideal strategy for creating timestamps used as strings
- Less performant than `Process.clock_gettime`
- Legacy strategy for duration and float timestamps

Other changes:
* Removed deprecated method, `Transaction.wrap`
* Remove unused time argument on `PipeService#shutdown`

# Related Github Issue
Closes: https://github.com/newrelic/newrelic-ruby-agent/issues/304
Closes: https://github.com/newrelic/newrelic-ruby-agent/issues/758

# Testing
The agent includes a suite of unit and functional tests which should be used to
verify your changes don't break existing functionality. These tests will run with 
Github Actions when a pull request is made. More details on running the tests locally can be found 
[here for our unit tests](https://github.com/newrelic/newrelic-ruby-agent/blob/main/test/README.md), 
and [here for our functional tests](https://github.com/newrelic/newrelic-ruby-agent/blob/main/test/multiverse/README.md).
For most contributions it is strongly recommended to add additional tests which
exercise your changes. 

# Reviewer Checklist
- [ ] Perform code review
- [ ] Add performance label
- [ ] Perform appropriate level of performance testing
- [ ] Confirm all checks passed
- [ ] Add version label prior to acceptance