@kaylareopelle 2021-08-27
### Performance Testing Results 

**TL;DR** The working branch completed the test run 12 seconds faster than the main branch

```
➜  performance git:(main) ✗ script/runner -B
65 tests, 0 failures, 576.01273 s total
Saved 65 results as baseline.
➜  performance git:(main) ✗ gco feature/use-monotonically-increasing-time
M       test/environments/lib/environments/runner.rb
Switched to branch 'feature/use-monotonically-increasing-time'
➜  performance git:(feature/use-monotonically-increasing-time) ✗ script/runner -C
65 tests, 0 failures, 562.130768 s total
+----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------+
| name                                                                             | before    | after     | delta  | allocs_before | allocs_after | allocs_delta | retained | retained_delta |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| ActiveRecordTest#test_helper_by_name                                             |   1.00 µs |   0.96 µs |  -3.7% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| ActiveRecordTest#test_helper_by_sql                                              |   2.66 µs |   2.41 µs |  -9.4% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| ActiveRecordSubscriberTest#test_subscriber_in_txn                                |   3.67 ms |   2.89 ms | -21.3% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| AgentAttributesTests#test_empty_agent_attributes                                 |   0.84 µs |   0.78 µs |  -7.2% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| AgentAttributesTests#test_with_wildcards                                         |   1.44 µs |   1.31 µs |  -8.9% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| AgentAttributesTests#test_with_tons_o_rules                                      |  33.92 µs |  31.33 µs |  -7.6% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| AgentAttributesTests#test_with_attribute_rules                                   |   1.63 µs |   1.19 µs | -27.2% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| AgentModuleTest#test_increment_metric_by_more_than_1                             |   5.46 µs |   4.04 µs | -26.0% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| AgentModuleTest#test_increment_metric_by_1                                       |   3.47 µs |   2.77 µs | -20.2% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| ConfigPerfTests#test_defaulting_access                                           |   0.09 µs |   0.08 µs | -18.3% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| ConfigPerfTests#test_missing_key                                                 |   0.10 µs |   0.08 µs | -27.2% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| ConfigPerfTests#test_blowing_cache                                               |   2.17 µs |   1.72 µs | -20.9% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| ConfigPerfTests#test_raw_access                                                  |   0.08 µs |   0.08 µs |  -2.5% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| DatastoresPerfTest#test_instrumentation_via_trace                                | 172.73 µs | 135.80 µs | -21.4% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| DatastoresPerfTest#test_wrap                                                     | 179.36 µs | 142.48 µs | -20.6% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| DatastoresPerfTest#test_notice_sql                                               |   2.54 µs |   1.99 µs | -21.7% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| DatastoresPerfTest#test_segment_notice_sql                                       |   1.09 µs |   0.97 µs | -10.8% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| ErrorCollectorTests#test_notice_error                                            | 203.50 µs | 175.07 µs | -14.0% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| ErrorCollectorTests#test_notice_error_with_custom_attributes                     | 480.10 µs | 178.67 µs | -62.8% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| ExternalSegment#test_external_cat_request                                        |   3.45 ms |   2.08 ms | -39.8% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| ExternalSegment#test_external_trace_context_request                              |   2.88 ms |   2.13 ms | -26.1% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| ExternalSegment#test_external_request                                            |   3.75 ms |   2.22 ms | -40.9% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| LoggingTest#test_logging                                                         |   8.07 µs |   7.02 µs | -13.0% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| Marshalling#test_basic_marshalling_json                                          |  35.57 ms |  29.15 ms | -18.1% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| Marshalling#test_json_marshalling_binary_strings                                 |  81.10 ms |  69.50 ms | -14.3% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| Marshalling#test_json_marshalling_utf16_strings                                  |  77.30 ms |  61.34 ms | -20.6% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| Marshalling#test_json_marshalling_latin1_strings                                 |  83.77 ms |  63.36 ms | -24.4% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| QueueTimePerfTests#test_queue_time_parsing                                       |  25.94 µs |  10.34 µs | -60.2% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| RackMiddleware#test_basic_middleware_stack                                       | 449.74 µs | 368.81 µs | -18.0% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| RackMiddleware#test_request_with_params_capture_params_off                       | 510.67 µs | 411.48 µs | -19.4% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| RackMiddleware#test_request_with_params_capture_params_on                        | 574.86 µs | 477.92 µs | -16.9% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| RedisTest#test_no_args                                                           |   0.45 µs |   0.40 µs | -10.4% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| RedisTest#test_args                                                              |   3.48 µs |   2.72 µs | -21.8% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| RedisTest#test_long_args                                                         |   4.98 µs |   3.75 µs | -24.6% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| RedisTest#test_pipelined                                                         |   8.10 µs |   7.55 µs |  -6.7% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| RulesEngineTests#test_rules_engine_init_transaction_rules                        |   0.92 µs |   0.80 µs | -13.2% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| RulesEngineTests#test_rules_engine_rename_transaction_rules                      |   1.65 µs |   1.38 µs | -16.5% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| RumAutoInsertion#test_rum_autoinsertion                                          | 462.23 µs | 396.03 µs | -14.3% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| RumAutoInsertion#test_rum_autoinsertion_with_x_ua_compatible                     | 282.27 µs | 271.10 µs |  -4.0% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| RumAutoInsertion#test_rum_autoinsertion_with_x_ua_compatible_after               | 466.42 µs | 387.21 µs | -17.0% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| SegmentTermsRuleTests#test_segment_terms_rule_apply                              |   5.09 µs |   5.37 µs |  +5.6% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| SegmentTermsRuleTests#test_segment_terms_rule_matches?                           |   1.07 µs |   1.13 µs |  +5.6% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| SqlObfuscationTests#test_obfuscate_cross_agent_tests                             |   1.96 ms |   1.95 ms |  -0.4% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| SqlObfuscationTests#test_obfuscate_sql                                           |  87.62 µs |  89.64 µs |  +2.3% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| SqlObfuscationTests#test_obfuscate_sql_postgres                                  |  79.57 µs |  74.82 µs |  -6.0% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| StartupShutdown#test_startup_shutdown                                            | 815.97 µs | 940.11 µs | +15.2% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| StatsHashPerfTest#test_merge                                                     | 273.35 µs | 269.29 µs |  -1.5% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| StatsHashPerfTest#test_record                                                    | 144.40 µs | 172.37 µs | +19.4% |               |              |         0.0% |        0 |           0.0% |
|----------------------------------------------------------------------------------+-----------+-----------+--------+---------------+--------------+--------------+----------+----------------|
| ThreadProfiling#test_gather_backtraces                                           |   1.37 ms |   1.38 ms |  +0.5% |               |              |         0.0% |        0 |           0.0% |
…